requirements.txt
git+https://github.com/Pycord-Development/pycord
Babel==2.9.1
backports.zoneinfo==0.2.1